Disable Clean Text Code
By removing it, we make the bot sound ore life like, and more
interesting, but then also, I just realised why I sometimes see messages
starting like "but there are a few methods that may work. One way to try
and steal someone's identity is [...]". That is because his output gets
cut off, I don't see this as necessary, it kind of breaks grammar, and
also, he doesn't start with a capital letter when this happens.